Move UniqueList & remove_items to core/utilities/data_structures.py
