+++ fedot/core/dag/graph_operator.py
@@ -1,8 +1,9 @@
 from copy import deepcopy
-from typing import Any, List, Optional, Union, Tuple, Container
+from typing import Any, List, Optional, Union, Tuple
 
 from fedot.core.dag.graph_node import GraphNode
 from fedot.core.pipelines.convert import graph_structure_as_nx_graph
+from fedot.core.utilities.data_structures import remove_items
 
 
 class GraphOperator:
@@ -28,10 +29,10 @@ class GraphOperator:
         """Delete node with all the parents it has.
         and delete all edges from removed nodes to remaining graph nodes."""
         subtree_nodes = node.ordered_subnodes_hierarchy()
-        remove_items(self._graph.nodes, subtree_nodes)
+        self._graph.nodes = remove_items(self._graph.nodes, subtree_nodes)
         # prune all edges coming from the removed subtree
         for node in self._graph.nodes:
-            remove_items(node.nodes_from, subtree_nodes)
+            node.nodes_from = remove_items(node.nodes_from, subtree_nodes)
 
     def update_node(self, old_node: GraphNode, new_node: GraphNode):
         self.actualise_old_node_children(old_node, new_node)
@@ -223,10 +224,3 @@ class GraphOperator:
                 for parent_node in node.nodes_from:
                     edges.append((parent_node, node))
         return edges
-
-
-def remove_items(collection: List, removed_items: Container):
-    """Removes all specified items from the list. Modifies original collection."""
-    if collection:
-        collection[:] = [item for item in collection if item not in removed_items]
-    return collection
