+++ fedot/core/optimisers/graph.py
@@ -5,7 +5,7 @@ from uuid import uuid4
 from fedot.core.dag.graph_node import GraphNode
 from fedot.core.dag.graph_operator import GraphOperator
 from fedot.core.dag.node_operator import NodeOperator
-from fedot.core.dag.unique_list import UniqueList
+from fedot.core.utilities.data_structures import UniqueList
 from fedot.core.log import Log, default_log
 from fedot.core.utils import DEFAULT_PARAMS_STUB
 from fedot.core.visualisation.graph_viz import GraphVisualiser
